pyproject.toml
[tool.poetry]
name = "sdr-api-adapter"
version = "0.1.0"
description = ""
authors = ["CrinitusFeles <crinitusfeles@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.108.0"
uvicorn = "^0.25.0"
loguru = "^0.7.2"
event = {git = "https://github.com/CrinitusFeles/Event.git"}
python-dotenv = "^1.0.0"
pyzmq = "^25.1.2"
gnuradio-pmt = {git = "https://github.com/CrinitusFeles/gnuradio-pmt-python.git"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
